+++ client/css/spotify-gdpr-chart.css
@@ -1,7 +1,7 @@
 #chart-container {
   position: relative;
 
-  height: 30rem;
+  height: 40rem;
 }
 
 .chart-scroll-container {
@@ -11,7 +11,7 @@
 }
 
 .chart {
-  height: 30rem;
+  height: 40rem;
   overflow: visible;
 
   background: transparent;
