+++ client/spotify-gdpr.html
@@ -28,6 +28,9 @@
       <div id="chart-container">
       </div>
 
+      <table id="table-container">
+      </table>
+
       <h2 class="header">How it works</h2>
       <p>
         The GDPR is a European law that requires tech companies to let you download the data that
@@ -58,6 +61,6 @@
     <script type="text/javascript" src="https://www.gstatic.com/charts/loader.js"></script>
 
     <script src="js/zip/zip.js"></script>
-    <script src="js/spotify-gdpr.js" type="module"></script>
+    <script src="js/gdpr/index.js" type="module"></script>
   </body>
 </html>
